Clean up CommentsWidget and convert it to TypeScript
